refactor: integrate loadGLTFWithResourcesFromPath for improved resource handling in renderGLTFToPNGBuffer
